+++ test/themes/themes_service_test.dart
@@ -150,7 +150,6 @@ void main() {
       final themesService = ThemesService();
 
       expect(themesService.colorNames(), [
-        'actionColor',
         'activeAudioControl',
         'alarm',
         'audioMeterBar',